requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
commonmark==0.9.1
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
dnspython==2.2.1
gunicorn==20.1.0
httpie==3.2.1
idna==3.4
loguru==0.6.0
multidict==6.0.4
pinecone-client==2.0.13
psycopg2==2.9.5 #psycopg2 might not work on mac so use psycopg2-binary==2.9.5
Pygments==2.14.0
PySocks==1.7.1
python-dateutil==2.8.2
pytube==12.1.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.10.1
rich==13.0.0
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
whitenoise==6.2.0
